stop pushing to develop and master
instead, make a single commit based on develop, after merging master in.

this way we can just continuously push to master, with the caveat that
final release blobs will *only* show up there, and never be
automatically merged into develop. the advantage of this is avoiding
race conditions with merging back on top of develop, as people push
while CI runs. it also makes the CI not inherently circular.

also, create final release tarballs, so we can put them in S3 and
propagate them to later stages of the pipeline easily.

Signed-off-by: Alex Suraci <asuraci@pivotal.io>