Update bosh dance instructions
The current bosh dance instructions have the unfortunate consequence that if you don't have a config/dev.yml and copy/paste it, Bosh will prompt you for the name of the release and receive the `bosh -n upload release` as the input. This fixes it so the prompt correctly waits for input.