Move the uaadb and ccdb properties to the properties.yml
This makes it easier to inherit as many properties as possible from the
databases property instead of duplicating them. With the exception of the
db_scheme in uaa which has to be set to `postgresql` as opposed to `postgres`
all the properties are now copied over from the databases block.

[finishes #133428661]