Make cellID configurable
Non-BOSH deployments may have a different mechanism to name VMs, so
make the cellID configurable via a property (together with the
advertise_domain it must form the FQDN of the cell).  If the property
is not set, then the default naming algorithm will still be used,
making this change completely backwards compatible.