examples/aws: use v2 BOSH CLI and bosh-deployment
This set of changes updates the examples/aws documentation and script
to use the new BOSH CLI and the bosh-deployment repo.

In addition to using the variable interpolation and generation features
of the BOSH CLI to generate the BOSH deployment manifest, these changes
also add ops-files to enable Datadog integration and to add a domain SAN
to the director TLS certificate.

There are some breaking changes to the arguments to the deploy script:

* An extra required argument specifies what action to take on the BOSH deployment (deploy or skip).
* The unused cf-release directory argument has been replaced with a path to the bosh-deployment repository.
* The stack name is now required.

By default, the v2 BOSH CLI is assumed to be on the PATH as `bosh`, but
the deploy script respects an alternate name or path in the `gobosh` environment variable.