Make max_cache_size_in_bytes configurable on windows cells
We need to lower the windows cells cache size to avoid running out of disk
space. According to the #greenhouse team, it is not possible to increase the
root disk size (on AWS). Currently windows is taking around 15G and the cache
is about 8G which leaves little room for apps to run.