Rework rep_windows drain script evacuate function to not swallow curl output
The combination of try/catch and $ErrorActionPreference = "Stop" causes
output of gocurl that is being redirected/piped to be lost. This change
introduces more manual error handling in the evacuate function so we can
ensure the output of gocurl is sent to the log file. This should
hopefully help debug deployments where an update to the rep fails due to
a failed drain script.

[finishes #169682305]