Manually add github.com/containers/image/go.mod and go.sum
gosub doesn't support go modules. It is not adding go.mod and go.sum to the
package spec.  That is the root cause of the compilation issues we are seeing
in CI [1].  We decided to add those two files manually to the package spec
file.  We could fix gosub to do the right thing in the future.

Example of the error that we see in CI:

+ go build -a -installsuffix static code.cloudfoundry.org/dockerapplifecycle/builder
src/github.com/containers/image/docker/docker_transport.go:8:2: cannot find package "github.com/containers/image/v5/docker/policyconfiguration" in any of:
	/var/vcap/data/packages/golang-1-linux/280153b3a8a27b088ef3d09212376a6c4f15b57b/src/github.com/containers/image/v5/docker/policyconfiguration (from $GOROOT)
	/var/vcap/data/compile/docker_app_lifecycle/src/github.com/containers/image/v5/docker/policyconfiguration (from $GOPATH)

[1]:
https://diego.ci.cf-app.com/teams/main/pipelines/main/jobs/warp-drive-deploy-cf-deployment/builds/2540#L5e7a5297:688:717

Authored-by: John Shahid <jshahid@pivotal.io>