docker-shell can start container without db
Allows `scripts/docker-shell` to easily start a standard container that
does *not* contain a DB. This can be useful for locally testing a CI
script if a `fly execute` workflow is not desired or possible.

For example, you could now invoke `DB=notset scripts/docker-shell` and
then invoke `scripts/ci/run-unit-tests-no-backing-store` to test that
script, which only runs tests that do not need a database.